+++ core/src/bitmap.rs
@@ -1,1 +1,2 @@
+pub mod bitmap_data;
 pub mod turbulence;
